medora_demo/requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyAudio==0.2.14
python-clerk==0.0.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
WTForms==3.2.1
